Move team-validator to sim/team-validator
