+++ sim/team-validator.js
